removing old files, ignoring generated files, some renaming
